pyproject.toml
[tool.poetry]
name = "darjeeling"
version = "0.2.0"
description = "Distributed, language-independent, compositional search-based program repair"
authors = ["Chris Timperley <ctimperley@cmu.edu>"]
license = "Apache-2.0"
repository = "https://github.com/squaresLab/Darjeeling"
readme = "README.rst"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.scripts]
darjeeling = 'darjeeling.cli:main'

[tool.poetry.dependencies]
python = ">=3.11,<4"
kaskara = { git = "https://github.com/ChrisTimperley/Kaskara.git", rev = "f8f49db12f04e5c4f482a1ea94b767b701710afe" }
attrs = ">=19.3.0"
bugzoo = ">=2.1"
comby = "^0.0.2"
pyroglyph = ">=0.1.0"
cement = "^3.0"
loguru = "^0.5"
requests = "2.31.0"
sourcelocation = ">=1.0.6,^1.0"
dockerblade = "^0.6,>=0.6.2"
websockets = "^10.0"
overrides = "^7.7.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
flake8-comprehensions = "^3.14.0"
flake8-future-annotations = "^1.1.0"
flake8-import-order = "^0.18.2"
flake8-quotes = "^3.3.2"
icecream = "^2.1.3"
mypy = "^1.7.0"
nox = "^2023.4.22"
pep8-naming = "^0.13.3"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
python = "^3.11"
ruff = "^0.1.6"
types-pyyaml = "*"

[tool.mypy]
ignore_missing_imports = false
strict = true

[[tool.mypy.overrides]]
module = "bugzoo.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cement.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyroglyph.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "comby.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]
select = [
    "ALL",  # includes all rules (including new ones)
]
per-file-ignores = {"*/__init__.py" = ["F401"]}
extend-ignore = [
    "ANN101",   # missing type annotation for 'self' in method
    "ANN102",   # missing type annotation for 'cls' in classmethod
    "B024",     # use of abstract base class without abstract methods
    "D100",     # missing docstring in public module
    "D101",     # missing docstring in public class
    "D102",     # missing docstring in public method
    "D103",     # missing docstring in public function
    "D104",     # missing docstring in public package
    "D105",     # missing docstring in magic method
    "D203",     # one-blank-line-before-class
    "D213",     # multi-line-summary-second-line
    "D401",     # first line of docstring should be in imperative mood
    "D413",     # missing blank line after last section
    "FIX001",   # unresolved FIXME
    "FIX002",   # TODO found
    "TD001",    # invalid TODO tag: `FIXME`
    "TD002",    # missing author in TODO
    "TD003",    # missing issue link on line following TODO
    "TD004",    # missing colon in TODO
    "TRY003",   # avoid specifying long messages outside the exception class
    "S101",     # use of assertions
    "SLF001",   # private member accessed
    "T201",     # use of print
    "D204",     # 1 blank line required after class docstring
    "C901",     # function is too complex (cyclomatic complexity)
    "PLR0912",  # too many branches
    "PGH003",   # use specific rule codes when ignoring type issues
    "FBT001",   # boolean typed positional argument in function definition
    "ARG001",   # unused function argument
    "PLR0913",  # too many arguments in function definition
    "PLR0911",  # too many return statements
    "PYI041",   # use `float` instead of `int | float`
    "D106",     # missing docstring in public nested class
    "S311",     # standard pseudo-random generators are not suitable for security/cryptographic purposes
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
